+++ src/FunFair.BuildCheck/ProjectChecks/MustNotHaveFxCopAnalyzerPackage.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.Logging;
+
+namespace FunFair.BuildCheck.ProjectChecks
+{
+    /// <summary>
+    ///     Checks that FX-Cop is not referenced for .net core 5.0 targets.
+    /// </summary>
+    public sealed class MustNotHaveFxCopAnalyzerPackage : MustNotReferencePackages
+    {
+        public MustNotHaveFxCopAnalyzerPackage(ILogger<MustNotHaveFxCopAnalyzerPackage> logger)
+            : base(new[] {"Microsoft.CodeAnalysis.FxCopAnalyzers"}, reason: ".net core 5 and later include this by default", logger: logger)
+        {
+        }
+    }
+}