+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ### Added
 - Check that project exists when using project references.
 - Check that a library does not depend on an executable.
+- Check that Microsoft.CodeAnalysis.FxCopAnalyzers is not enabled for .net 5.0 targets
 ### Fixed
 ### Changed
 - FF-1429 - Updated Microsoft.VisualStudio.Threading.Analyzers to 16.8.55